+++ packages/@pufflig/ps-sdk/src/index.ts
@@ -1,4 +1,4 @@
-export { createCompletion } from "./createCompletion";
-export { streamCompletion } from "./streamCompletion";
-export { runDeployment } from "./runDeployment";
+export { createCompletion } from "./completionAPI/createCompletion";
+export { streamCompletion } from "./completionAPI/streamCompletion";
+export { runDeployment } from "./deploymentAPI/runDeployment";
 export { type Completion, type CreateCompletionInput } from "./types";
